requirements.txt
tgcrypto
asyncio
ffmpeg-python
py-tgcalls
git+https://github.com/pyrogram/pyrogram@master
git+https://github.com/UNREALBOT12/yt-dll@master
youtube-search-python==1.4.9
youtube-search
httpx==0.13.3
python-dotenv
gitpython
aiofiles
aiohttp
requests
heroku3
psutil
future
wget
